+++ cmp/grid/filter/GridFilterFieldSpec.ts
@@ -7,12 +7,12 @@
 import {ColumnRenderer} from '@xh/hoist/cmp/grid';
 import {HoistInputProps} from '@xh/hoist/cmp/input';
 import {PlainObject} from '@xh/hoist/core';
-import {FieldFilterOperator, parseFilter, StoreRecord, View} from '@xh/hoist/data';
+import {FieldFilterOperator, Filter, parseFilter} from '@xh/hoist/data';
 import {
     BaseFilterFieldSpec,
     BaseFilterFieldSpecConfig
 } from '@xh/hoist/data/filter/BaseFilterFieldSpec';
-import {castArray, compact, flatten, isDate, isEmpty, uniqBy} from 'lodash';
+import {castArray, compact, flatMap, isDate, isEmpty, uniqBy} from 'lodash';
 import {GridFilterModel} from './GridFilterModel';
 
 export interface GridFilterFieldSpecConfig extends BaseFilterFieldSpecConfig {
@@ -71,68 +71,61 @@ export class GridFilterFieldSpec extends BaseFilterFieldSpec {
     //------------------------
     loadValuesFromSource() {
         const {filterModel, field, source, sourceField} = this,
-            columnFilters = filterModel.getColumnFilters(field),
-            sourceStore = source instanceof View ? source.cube.store : source,
-            allRecords = sourceStore.allRecords;
+            columnFilters = filterModel.getColumnFilters(field);
 
-        // Apply external filters *not* pertaining to this field to the sourceStore
-        // to get the filtered set of available values to offer as options.
-        const cleanedFilter = this.cleanFilter(filterModel.filter);
-        let filteredRecords = allRecords;
-        if (cleanedFilter) {
-            const testFn = parseFilter(cleanedFilter).getTestFn(sourceStore);
-            filteredRecords = allRecords.filter(testFn);
-        }
+        // All possible values for this field from our source.
+        const allSrcVals = source.getValuesForFieldFilter(field).map(it => this.toDisplayValue(it));
 
-        // Get values from current column filter
-        const filterValues = [];
-        columnFilters.forEach(filter => {
-            const newValues = castArray(filter.value).map(value => {
-                value = sourceField.parseVal(value);
-                return filterModel.toDisplayValue(value);
-            });
-            filterValues.push(...newValues);
-        });
+        // Values from current column filter.
+        const colFilterVals = flatMap(columnFilters, filter => {
+            return castArray(filter.value).map(val => sourceField.parseVal(val));
+        }).map(it => this.toDisplayValue(it));
 
-        // Combine unique values from record sets and column filters.
-        const allValues = uniqBy(
-            [...flatten(allRecords.map(rec => this.valueFromRecord(rec))), ...filterValues],
-            this.getUniqueValue
-        );
-        let values;
-        if (cleanedFilter) {
-            values = uniqBy(
-                [
-                    ...flatten(filteredRecords.map(rec => this.valueFromRecord(rec))),
-                    ...filterValues
-                ],
-                this.getUniqueValue
-            );
+        // Combine + unique - these are all values that *could* be shown in the filter UI.
+        const allValues = uniqBy([...allSrcVals, ...colFilterVals], this.getUniqueValue);
+
+        // Create a filter with other filters *not* pertaining to this field, if any.
+        // This filter will be used to get the subset of all possible values for this field that
+        // exist in records passing those other filters.
+        const otherFieldsFilter = this.cleanFilter(filterModel.filter);
+
+        let values: any[];
+
+        if (otherFieldsFilter) {
+            // If we have filters on other fields, get values from source that pass that filter.
+            // These will be the set of values shown in the filter UI.
+            const filteredSrcVals = source
+                .getValuesForFieldFilter(field, otherFieldsFilter)
+                .map(it => this.toDisplayValue(it));
+            values = uniqBy([...filteredSrcVals, ...colFilterVals], this.getUniqueValue);
         } else {
+            // Otherwise, all possible values are shown.
             values = allValues;
         }
 
         this.values = values.sort();
+
+        // ValueCount always reflects total possible values for this field, allowing the UI
+        // to indicate to the user if some values are hidden due to other active filters.
         this.valueCount = allValues.length;
     }
 
     // Recursively modify a Filter|CompoundFilter to remove all FieldFilters referencing this column
-    private cleanFilter(filter) {
-        if (!filter) return filter;
-
-        const {field, filters, op} = filter;
-        if (filters) {
+    private cleanFilter(filter: Filter): Filter {
+        if (Filter.isCompoundFilter(filter)) {
+            const {filters, op} = filter;
             const ret = compact(filters.map(it => this.cleanFilter(it)));
-            return !isEmpty(ret) ? {op, filters: ret} : null;
-        } else if (field === this.field) {
+            return !isEmpty(ret) ? parseFilter({op, filters: ret}) : null;
+        }
+
+        if (Filter.isFieldFilter(filter) && filter.field === this.field) {
             return null;
         }
 
         return filter;
     }
 
-    private valueFromRecord(record: StoreRecord) {
-        const {filterModel, field} = this;
-        return filterModel.toDisplayValue(record.get(field));
+    private toDisplayValue(val: any): any {
+        return this.filterModel.toDisplayValue(val);
     }
 }